Finalize new interpretability code
