+++ src/drvi/scvi_tools_based/model/base/__init__.py
@@ -1,7 +1,9 @@
 from ._archesmixin import DRVIArchesMixin
 from ._generative_mixin import GenerativeMixin
+from ._interpretability_mixin import InterpretabilityMixin
 
 __all__ = [
     "DRVIArchesMixin",
     "GenerativeMixin",
+    "InterpretabilityMixin",
 ]
